Finished (again) the exercices about Lists and Sorting Lists Challenge.
I had a problem in a Sorting program, tried to sort with: people = people.OrderBy(o=>o.FullName).ToList(); but is no working and I dont know why.
After this I taked the answer from the exercices and this way its working, they are giving me the same answer.
Done a foreach loop for see the List, there was ordered but just passed the test with the method .sort((x, y)).