We added 3 new programs to our GitHub.
One explains about DataTime methods and functions, the next one works with TimeSpan, we added some examples to them.

Finally we worked with a DataTime challengue, I tried to discover about why is not appearing the PM/AM format when I declare it on the method toString("hh:MM:ss tt").
It supposed the "tt" text in the method toString makes appear the format 12 hours clock or 24h.
At least the program works fine, and works, but still not appearing the time format.