Added an exercice about a Windows Form Challenge.
The program has 1 listBox where you can see the to do things, it has a button for edit and delete a task, and another one for add a new one.

It works with a priority, you can Up and Down a task if you got more priority about something. Like me with that exercices.

If you double click on a task the status of the task (true/false) changes.

Everytime you done a task, and you delete it, the program reorganice the Element index again, and you can move from Index 0 to whatever you want, but no another Indexes.

You can use your arrow keys (Up/Down) if you got clicked a task for order it.