Added a new exercice what does a replacement string.
We completed the challenge. Added a variable for use our filePath easily, created a list what has de content of the document because its readed.

The console ask for the text what you want to replace, and later what word to put in the replaced word.

Used the method Replace with the old text, and the new text and finally we wrote it in the doc.